fix(typecheck): resolve all 22 mypy strict errors
graph_builder.py:
  - cast() return of _coerce_enum in _parse_generation (no-any-return)
  - convert lcom4 float→int at ClassNode construction site (arg-type)
  - _embed_parameter signature: dict[str,Any] → ParameterFact (arg-type)
  - import ParameterFact from parser.models, add cast to imports

graph_assembler.py:
  - remove 15 stale type:ignore[union-attr] comments (unused-ignore)

parser/models.py:
  - lcom4: float|None → int|None (LCOM4 is always an integer count)

input/source_discoverer.py:
  - PathSpec → PathSpec[Any] with Any import (type-arg)

config/yaml_source.py:
  - get_field_value override: (str,Any)→tuple → (FieldInfo,str)→Any (override)
  - import FieldInfo from pydantic.fields

pyproject.toml:
  - add types-PyYAML to dev deps
  - disable_error_code = [import-untyped] for yaml (no stubs in all envs)

+++ codeograph/parser/models.py
@@ -201,7 +201,7 @@ class ParsedFile(_ParsedFileBase, total=False):
     entry_point: bool
     wmc: int | None  # Weighted Methods per Class; M7
     cbo: int | None  # Coupling Between Objects; M7
-    lcom4: float | None  # Lack of Cohesion of Methods 4; M7
+    lcom4: int | None  # Lack of Cohesion of Methods 4; M7
 
     # --- interface-only ---
     extends_interfaces: list[str]
